# Somewhat improve sprite processing