--- a/Robust.Client/Graphics/Clyde/Clyde.Rendering.cs
+++ b/Robust.Client/Graphics/Clyde/Clyde.Rendering.cs
@@ -1008,45 +1008,6 @@ namespace Robust.Client.Graphics.Clyde
             PointList,
         }
 
-        private sealed class SpriteDrawingOrderComparer : IComparer<int>
-        {
-            private readonly RefList<(SpriteComponent, Vector2, Angle, Box2)> _drawList;
-
-            public SpriteDrawingOrderComparer(RefList<(SpriteComponent, Vector2, Angle, Box2)> drawList)
-            {
-                _drawList = drawList;
-            }
-
-            public int Compare(int x, int y)
-            {
-                var a = _drawList[x];
-                var b = _drawList[y];
-
-                var cmp = a.Item1.DrawDepth.CompareTo(b.Item1.DrawDepth);
-                if (cmp != 0)
-                {
-                    return cmp;
-                }
-
-                cmp = a.Item1.RenderOrder.CompareTo(b.Item1.RenderOrder);
-
-                if (cmp != 0)
-                {
-                    return cmp;
-                }
-
-                // compare the top of the sprite's BB for y-sorting. Because screen coordinates are flipped, the "top" of the BB is actually the "bottom".
-                cmp = a.Item4.Top.CompareTo(b.Item4.Top);
-
-                if (cmp != 0)
-                {
-                    return cmp;
-                }
-
-                return a.Item1.Owner.CompareTo(b.Item1.Owner);
-            }
-        }
-
         private readonly struct FullStoredRendererState
         {
             public readonly Matrix3 ProjMatrix;