This PR improves the pre-rendering sprite processing (getting world position, rotation, and bounding boxes). It also moves most of the processing & sorting logic into a separate partial class. AFAICT this is generally not a bottleneck in most situations, but can becomes problematic if there are many sprites on screen (e.g., when zooming out). 

I just tested this using a basic test map with ~15k sprites on screen with random positions. On master release mode I get about 67fps, with a running average of 4.9ms spent inside `ProcessSpriteEntities()`. With these changes, I get ~90fps with 0.67ms spent processing sprites. About half of the speedup just comes from faster transform & bounding box calculations (4.9ms -> 2.5ms), the rest comes from parallelising (2.5ms -> 0.62ms, though it actually increases to ~3.3ms if only using a single thread).

When it comes to SS14 gameplay, when moving around box station (400-500 sprites on screen), my desktop & laptop framerate doesn't actually change all that much. But maybe this will help potato computers?