@ElectroJr 2022-12-19
> When it comes to SS14 gameplay, when moving around box station (400-500 sprites on screen), my desktop & laptop framerate doesn't actually change all that much. But maybe this will help potato computers?

![image](https://cdn.discordapp.com/attachments/900473545040207942/1054339097583431750/image.png)

Well never mind about 15k being an unreasonable test number. If botany is often this active the sprite count might actually approach 10k every now and then, though that's also a performance issue that needs fixing for many other reasons.