@PJB3005 2022-12-19
> If nothing else this will likely help OD where it's not unusual to have a few thousand sprites on screen at a time.

I don't think any of the code here is hit by OpenDream