@ElectroJr 2022-12-21
Currently there is a bug somewhere that results in transform matrices not being properly set during rendering.
#3586 fixed it by just ensuring matrices were always updated, but apparently this broke physics.
So I guess DNM until thats fixed somehow or #3479 just gets rid of the deferral